mediaProperties: use wrapper classes instead of structs of copies
* Remove:
     * file mediaProperty (hpp/cpp)
     * files <type>StreamProperties (hpp)
* Add:
     * file <type>Properties (hpp/cpp)
     * util (hpp/cpp)
* Rename:
     * file printMediaProperty.hpp to print.hpp
     * struct Properties to class FileProperties

+++ src/AvTranscoder/transform/AudioTransform.cpp
@@ -84,7 +84,7 @@ void AudioTransform::convert( const Frame& srcFrame, Frame& dstFrame )
 		throw std::runtime_error( "unable to convert audio samples" );
 	}
 
-	int nbOutputSamples = nbOutputSamplesPerChannel * static_cast<const AudioFrame&>( dstFrame ).desc().getChannels();
+	size_t nbOutputSamples = nbOutputSamplesPerChannel * static_cast<const AudioFrame&>( dstFrame ).desc().getChannels();
 	
 	if( dstFrame.getSize() != nbOutputSamples )
 		dstFrame.getBuffer().resize( nbOutputSamples, 0 );
